requirements.txt
# Core dependencies
psycopg2-binary==2.9.5
sqlalchemy==2.0.16
pandas==2.0.2
numpy==1.24.3

# Data manipulation and analysis
jupyter==1.0.0
matplotlib==3.7.1
seaborn==0.12.2

# Documentation
sphinx==6.1.3
sphinx-rtd-theme==1.2.0

# Testing
pytest==7.3.1
pytest-cov==4.1.0

# Development tools
black==23.3.0
flake8==6.0.0
isort==5.12.0

# Docker integration
docker==6.1.1